+++ VersionHistory.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 6.7.0
+
+- Added `allowBackgroundScrolling` feature to `Modal` and `ShareDialog`.
+
+## 6.6.6
+
+- Fix crash in Safari with number inputs #494
+
 ## 6.6.5
 
 - Republish